Update Metric_Evaluation_sleap_mw.ipynb
Really nice job! I think there are two things from your script that would be immediately great for our error outputs -- kernel densities of velocity, as well as velocity flag for potentially problematic frames. I will continue to format that markdown script I sent you into something a little more plug-and-play that includes my old visualizations. Most of these are identical to yours (ie, histograms of scores). I also incorporated a count of cells with missing data, to try and get a sense of when our tracker was losing fish. This needs to be worked on. I will get back to you, but it may be helpful for us to have more of a sit down meeting and actually go through and compare scripts and piece together what we have. LMK.